requirements.txt
Flask>=2.0.0
Flask-Cors>=3.0.0
python-dotenv>=0.19.0
requests>=2.25.0
psutil>=5.8.0
PyJWT>=2.0.0
bcrypt>=3.2.0
cachetools>=5.0.0